demo002/src/test/java/com/example/demo/AppControllerTest.java
package com.example.demo;

import org.apache.kafka.clients.consumer.ConsumerConfig;
import org.apache.kafka.clients.consumer.ConsumerRecord;
import org.apache.kafka.clients.producer.ProducerConfig;
import org.apache.kafka.common.serialization.IntegerDeserializer;
import org.apache.kafka.common.serialization.IntegerSerializer;
import org.apache.kafka.common.serialization.StringDeserializer;
import org.apache.kafka.common.serialization.StringSerializer;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.context.ApplicationContext;
import org.springframework.kafka.core.DefaultKafkaConsumerFactory;
import org.springframework.kafka.core.DefaultKafkaProducerFactory;
import org.springframework.kafka.core.KafkaTemplate;
import org.springframework.kafka.core.ProducerFactory;
import org.springframework.kafka.listener.KafkaMessageListenerContainer;
import org.springframework.kafka.listener.MessageListener;
import org.springframework.kafka.listener.config.ContainerProperties;
import org.springframework.test.context.ContextConfiguration;
import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
import org.springframework.test.context.junit4.SpringRunner;

import java.util.HashMap;
import java.util.List;
import java.util.Map;
import java.util.concurrent.CountDownLatch;
import java.util.concurrent.TimeUnit;

import static org.junit.Assert.*;

@RunWith(SpringRunner.class)
@SpringBootTest()
public class AppControllerTest {

    @Autowired
    private AppController controller;

    @Test
    public void getEmployees() {
        List<String> emp = controller.getEmployees();
        assertEquals(1,((List) emp).size());
        assertEquals("BB",((List) emp).get(0));

    }

    @Test
    public void getEmployees2() {
        String emp = controller.getEmployees2();
        assertEquals("AACC",emp);
    }

    @Test
    public void putAMessageIntoTopic() {
        String x = null;
        try {
            x = controller.putAMessageIntoTopic();
        } catch (Exception e) {
            assert(true);
        }
        assertEquals("OK",x);

    }

}